decluttering config.c (rules go into laws.c)
